+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+- Add official support for Django 5.1
+- Add official support Python 3.13
+- Add support for Django 5.1's `querystring` template tag
 - Drop official support for deprecated Python 3.7
 
 ## 1.3.2 - 2024-06-05
@@ -19,7 +22,7 @@
 - Add official support for Django 3.0 to 5.0 with:
     - Addition of automated test coverage
     - Review of all the built-in template tags and filters
-- Add support for Django 5's `escapeseq` filter
+- Add support for Django 5's `escapeseq` template filter
 - Use [tox](https://tox.wiki/) to manage test suite environment
 
 ## 1.2.0 - 2024-04-18